foreign key is working

+++ database/migrations/2023_01_16_132001_create_classe_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('classe', function (Blueprint $table) {
             $table->id();
-            $table->integer('id_etablissement');
-            $table->integer('id_referenciel');
+            $table->unsignedBigInteger('id_etablissement');
+            $table->unsignedBigInteger('id_referenciel');
             $table->string('libelle');
             $table->string('niveau');
             $table->integer('effectif');
