+++ database/migrations/2023_01_16_132754_create_groupe_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('groupe', function (Blueprint $table) {
             $table->id();
-            $table->integer('id_matiere');
-            $table->integer('id_enseignant');
+            $table->unsignedBigInteger('id_matiere');
+            $table->unsignedBigInteger('id_enseignant');
             $table->timestamps();
         });
     }
